move G_SpawnFire from sg_utils into sg_missiles to get rid of another CBSE.h include
